+++ CHANGES.rst
@@ -26,6 +26,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- Provide better GeoIP accuracy numbers for cities and subdivisions.
+
 - Fix cell queries containing invalid area codes but valid cids.
 
 - #242: Add WiFi stats to region specific stats page.